update api interface

+++ spring-boot/webflux-jpa/src/main/java/org/hung/web/PersonController.java
@@ -1,10 +1,14 @@
 package org.hung.web;
 
+import java.util.Optional;
 import java.util.UUID;
 
 import org.hung.pojo.Person;
 import org.hung.repo.PersonRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,14 +16,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 @Slf4j
-@RestController
+@Controller
 @RequestMapping("/api/v1/person")
 public class PersonController {
 
@@ -27,33 +30,46 @@ public class PersonController {
 	private PersonRepository repo;
 	
 	@GetMapping("/{id}")
-	public Mono<Person> findById(@PathVariable String id) {
-		return Mono.justOrEmpty(repo.findById(id));
-		//return Mono.fromCallable(() -> repo.findById(id).get());
+	public ResponseEntity<Mono<Person>> findById(@PathVariable String id) {
+		Optional<Person> result = repo.findById(id);
+		if (result.isEmpty()) {
+			return ResponseEntity.notFound().build();
+		} else {
+			return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(Mono.justOrEmpty(result));
+		}
 	}
 	
 	@GetMapping("/page")
-	public Flux<Person> findByPage() {
-		return Flux.fromIterable(repo.findAll());
+	public ResponseEntity<Flux<Person>> findByPage() {
+		return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(Flux.fromIterable(repo.findAll()));
 	}
 	
 	@PostMapping("/")
-	public Mono<Person> create(@RequestBody Person person) {
+	public ResponseEntity<Mono<Person>> create(@RequestBody Person person) {
 		person.setId(UUID.randomUUID().toString());
 		Person result = repo.save(person);
-		return Mono.just(result);
+		return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(Mono.just(result));
 	}
 	
 	@PutMapping("/{id}")
-	public Mono<Person> update(@PathVariable String id, @RequestBody Person person) {
-		Person result = repo.save(person);
-		return Mono.just(result);
+	public ResponseEntity<Mono<Person>> update(@PathVariable String id, @RequestBody Person person) {
+		Optional<Person> found = repo.findById(id);
+		if (found.isEmpty()) {
+			return ResponseEntity.notFound().build();
+		} else {
+			Person result = repo.save(person);
+			return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(Mono.just(result));
+		}
 	}
 	
 	@DeleteMapping("/{id}")
-	public Mono<Person> delete(@PathVariable String id) {
-		Mono<Person> result = Mono.justOrEmpty(repo.findById(id));
-		repo.deleteById(id);
-		return result;
+	public ResponseEntity<Mono<Person>> delete(@PathVariable String id) {
+		Optional<Person> found = repo.findById(id);
+		if (found.isEmpty()) {
+			return ResponseEntity.notFound().build();
+		} else {
+			repo.deleteById(id);
+			return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(Mono.justOrEmpty(found));
+		}
 	}
 }
